Use local eirini via vendoring dependencies
As of today, eirini-controller uses code from eirini. For development
purposes, it is convenient to be able to build the controller with the
local eirini version. The code from eirini is howver not avaialbe on the
docker builder image (and cannot be mounted as it is outside the builder
context), that's why this commit introduces a `replace` statement in
go.mod thus making eirini avaialble in the vendor directory

Issue: cloudfoundry-incubator/eirini-release#217